Create template-from-json-ingrid.html
ok -  here's a minimalist but not dull (it has cats!)  example of the current method of making products from  JSON data with alpines templating bits - hopefully. I've made useful comments to show not only how it works but give viable examples of a few different methods for displaying data, in an understandable context.